requirements.txt
playwright 
chromadb 
tiktoken 
streamlit
google-generativeai 